refactor(wallet): consolidate admin and user wallet repositories
Remove IAdminWalletRepository and merge its functionality into IWalletRepository
Update all admin use cases to use the consolidated repository
Remove AdminWalletRepositoryImpl and move its logic to WalletRepositoryImpl
Update DI configuration to reflect repository consolidation
